feat: pop-up-bar moved to content-panels
