+++ src/components/content-panels/pop-up-bar.ts
